// apps/lab/src/routes/graphql/-sub/07-optimistic-updates/__generated__/solutionDemoOptimisticQuery.graphql.ts
export type solutionDemoOptimisticQuery$variables = {
  postId: string;
};
export type solutionDemoOptimisticQuery$data = {
  readonly post: {
    readonly comments: ReadonlyArray<{
      readonly author: {
        readonly name: string;
      };
      readonly body: string;
      readonly id: string;
    }>;
    readonly id: string;
  } | null | undefined;
};
export type solutionDemoOptimisticQuery = {
  response: solutionDemoOptimisticQuery$data;
  variables: solutionDemoOptimisticQuery$variables;
};